Working on user section
This includes the ability to disable/delete & enable/restore a user. Users are now implementing the softdeletes feature to ensure that user data can be recovered if need be (Complete deletion process to come). Added seeders to help with permissions and roles, currently just generating the bare minimum names for roles and permissions. There are no relationships being drawn between these two models as of yet. Finally Added the navbar option, and there is more to come with that as well.